Move clang helper functions to clang.py, rename
